relocate some things
